dvm_r2/requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-allauth==0.60.0
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-import-export==3.3.6
django-mathfilters==1.0.0
django-tailwind==3.8.0
et-xmlfile==1.1.0
idna==3.6
Jinja2==3.1.2
mailjet-rest==1.3.4
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.4
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
tzdata==2023.4
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
